+++ lib/shared/src/sourcegraph-api/completions/nodeClient.ts
@@ -45,7 +45,7 @@ export class SourcegraphNodeCompletionsClient extends SourcegraphCompletionsClie
                 // If the request failed with a rate limit error, wraps the
                 // error in RateLimitError.
                 function handleError(e: Error): void {
-                    log?.onError(e.message)
+                    log?.onError(e.message, e)
 
                     if (res.statusCode === 429) {
                         // Check for explicit false, because if the header is not set, there
@@ -132,7 +132,7 @@ export class SourcegraphNodeCompletionsClient extends SourcegraphCompletionsClie
                     'Could not connect to Cody. Please ensure that Cody app is running or that you are connected to the Sourcegraph server.'
                 )
             }
-            log?.onError(error.message)
+            log?.onError(error.message, e)
             cb.onError(error)
         })